// src/modules/invoice/usecase/generate-invoice/generate-invoice.dto.ts
export interface GenerateInvoiceInputDto {
    id?: string;
    name: string;
    document: string;
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
    items: {
        id: string,
        name: string,
        price: number
    }[];
}

export interface GenerateInvoiceOutputDto {
    id: string;
    name: string;
    document: string;
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
    items: {
        id: string,
        name: string,
        price: number
    }[];
    total: number;
}